+++ Sources/SwiftDisc/Models/Guild.swift
@@ -3,4 +3,6 @@ import Foundation
 public struct Guild: Codable, Hashable {
     public let id: Snowflake
     public let name: String
+    public let owner_id: Snowflake?
+    public let member_count: Int?
 }