+++ Sources/SwiftDisc/Models/GuildBan.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct GuildBan: Codable, Hashable {
+    public let reason: String?
+    public let user: User
+}