+++ Sources/SwiftDisc/Models/GuildWidgetSettings.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct GuildWidgetSettings: Codable, Hashable {
+    public let enabled: Bool
+    public let channel_id: Snowflake?
+}